# contrib/google.golang.org/grpc: Cache SpanType option (-2 allocs)