The SpanType option is set to a constant value for gRPC spans. Cache
a single instance of this option to save two allocations per call for
all gRPC server and client spans. The two allocations are the result
of tracer.Tag, and then the convTString operation required to convert
the string argument to interface{}.

The microbenchmark shows that this does reduce CPU time, but within
the "noise" on my laptop. However, the bytes and number of allocation
clearly decreases. This seems like a small and low risk win.

```
name                                                         old time/op    new time/op    delta
UnaryServerInterceptor/ok_no_metadata-8                        5.80µs ±52%    5.11µs ±13%    ~     (p=0.912 n=10+10)
UnaryServerInterceptor/ok_with_metadata_no_parent-8            6.79µs ±76%    5.46µs ±11%    ~     (p=0.436 n=10+10)
UnaryServerInterceptor/ok_with_metadata_with_parent-8          5.20µs ±47%    4.57µs ± 9%    ~     (p=0.489 n=9+9)
UnaryServerInterceptor/ok_no_metadata_with_analytics_rate-8    5.57µs ±32%    5.01µs ±16%    ~     (p=0.258 n=9+9)
UnaryServerInterceptor/error_no_metadata-8                     14.2µs ±22%    13.3µs ±12%    ~     (p=0.863 n=9+9)
UnaryServerInterceptor/error_no_metadata_no_stack-8            7.41µs ±31%    6.72µs ±10%    ~     (p=0.931 n=9+9)

name                                                         old alloc/op   new alloc/op   delta
UnaryServerInterceptor/ok_no_metadata-8                        4.82kB ± 1%    4.75kB ± 0%  -1.33%  (p=0.000 n=10+10)
UnaryServerInterceptor/ok_with_metadata_no_parent-8            5.04kB ± 0%    4.97kB ± 0%  -1.37%  (p=0.000 n=8+10)
UnaryServerInterceptor/ok_with_metadata_with_parent-8          4.14kB ± 1%    4.07kB ± 0%  -1.50%  (p=0.000 n=10+9)
UnaryServerInterceptor/ok_no_metadata_with_analytics_rate-8    5.03kB ± 1%    4.96kB ± 1%  -1.40%  (p=0.000 n=10+10)
UnaryServerInterceptor/error_no_metadata-8                     11.4kB ± 0%    11.3kB ± 0%  -0.48%  (p=0.000 n=8+10)
UnaryServerInterceptor/error_no_metadata_no_stack-8            6.34kB ± 1%    6.26kB ± 1%  -1.19%  (p=0.000 n=10+10)

name                                                         old allocs/op  new allocs/op  delta
UnaryServerInterceptor/ok_no_metadata-8                          46.0 ± 0%      44.0 ± 0%  -4.35%  (p=0.000 n=10+10)
UnaryServerInterceptor/ok_with_metadata_no_parent-8              49.0 ± 0%      47.0 ± 0%  -4.08%  (p=0.000 n=10+10)
UnaryServerInterceptor/ok_with_metadata_with_parent-8            40.0 ± 0%      38.0 ± 0%  -5.00%  (p=0.000 n=10+10)
UnaryServerInterceptor/ok_no_metadata_with_analytics_rate-8      48.0 ± 0%      46.0 ± 0%  -4.17%  (p=0.000 n=10+10)
UnaryServerInterceptor/error_no_metadata-8                       69.0 ± 0%      67.0 ± 0%  -2.90%  (p=0.000 n=10+10)
UnaryServerInterceptor/error_no_metadata_no_stack-8              55.0 ± 0%      53.0 ± 0%  -3.64%  (p=0.000 n=10+10)
```